+++ libevmasm/Assembly.cpp
@@ -117,69 +117,36 @@ string Assembly::locationFromSources(StringMap const& _sourceCodes, SourceLocati
 
 ostream& Assembly::streamAsm(ostream& _out, string const& _prefix, StringMap const& _sourceCodes) const
 {
-	_out << _prefix << ".code:" << endl;
-	for (AssemblyItem const& i: m_items)
+	for (size_t i = 0; i < m_items.size(); ++i)
 	{
-		_out << _prefix;
-		switch (i.type())
+		AssemblyItem const& item = m_items[i];
+		if (!item.location().isEmpty() && (i == 0 || m_items[i - 1].location() != item.location()))
 		{
-		case Operation:
-			_out << "  " << instructionInfo(i.instruction()).name  << "\t" << i.getJumpTypeAsString();
-			break;
-		case Push:
-			_out << "  PUSH" << dec << max<unsigned>(1, dev::bytesRequired(i.data())) << " 0x" << hex << i.data();
-			break;
-		case PushString:
-			_out << "  PUSH \"" << m_strings.at((h256)i.data()) << "\"";
-			break;
-		case PushTag:
-			if (i.data() == 0)
-				_out << "  PUSH [ErrorTag]";
-			else
-			{
-				size_t subId = i.splitForeignPushTag().first;
-				if (subId == size_t(-1))
-					_out << "  PUSH [tag" << dec << i.splitForeignPushTag().second << "]";
-				else
-					_out << "  PUSH [tag" << dec << subId << ":" << i.splitForeignPushTag().second << "]";
-			}
-			break;
-		case PushSub:
-			_out << "  PUSH [$" << size_t(i.data()) << "]";
-			break;
-		case PushSubSize:
-			_out << "  PUSH #[$" << size_t(i.data()) << "]";
-			break;
-		case PushProgramSize:
-			_out << "  PUSHSIZE";
-			break;
-		case PushLibraryAddress:
-			_out << "  PUSHLIB \"" << m_libraries.at(h256(i.data())) << "\"";
-			break;
-		case Tag:
-			_out << "tag" << dec << i.data() << ": " << endl << _prefix << "  JUMPDEST";
-			break;
-		case PushData:
-			_out << "  PUSH [" << hex << (unsigned)i.data() << "]";
-			break;
-		default:
-			BOOST_THROW_EXCEPTION(InvalidOpcode());
+			_out << _prefix << "    /*";
+			if (item.location().sourceName)
+				_out << " \"" + *item.location().sourceName + "\"";
+			if (!item.location().isEmpty())
+				_out << ":" << to_string(item.location().start) + ":" + to_string(item.location().end);
+			_out << " */" << endl;
 		}
-		_out << "\t\t" << locationFromSources(_sourceCodes, i.location()) << endl;
+		_out << _prefix << (item.type() == Tag ? "" : "  ") << item.toAssemblyText() << endl;
 	}
 
 	if (!m_data.empty() || !m_subs.empty())
 	{
-		_out << _prefix << ".data:" << endl;
+		_out << _prefix << "stop" << endl;
+		Json::Value data;
 		for (auto const& i: m_data)
-			if (u256(i.first) >= m_subs.size())
-				_out << _prefix << "  " << hex << (unsigned)(u256)i.first << ": " << dev::toHex(i.second) << endl;
+			assertThrow(u256(i.first) < m_subs.size(), AssemblyException, "Data not yet implemented.");
+
 		for (size_t i = 0; i < m_subs.size(); ++i)
 		{
-			_out << _prefix << "  " << hex << i << ": " << endl;
-			m_subs[i]->stream(_out, _prefix + "  ", _sourceCodes);
+			_out << endl << _prefix << "sub_" << i << ": assembly {\n";
+			m_subs[i]->streamAsm(_out, _prefix + "    ", _sourceCodes);
+			_out << _prefix << "}" << endl;
 		}
 	}
+
 	return _out;
 }
 
@@ -449,7 +416,7 @@ LinkerObject const& Assembly::assemble() const
 		switch (i.type())
 		{
 		case Operation:
-			ret.bytecode.push_back((byte)i.data());
+			ret.bytecode.push_back((byte)i.instruction());
 			break;
 		case PushString:
 		{