+++ libevmasm/AssemblyItem.cpp
@@ -20,6 +20,7 @@
  */
 
 #include "AssemblyItem.h"
+#include <libevmasm/SemanticInformation.h>
 #include <fstream>
 
 using namespace std;
@@ -28,19 +29,19 @@ using namespace dev::eth;
 
 AssemblyItem AssemblyItem::toSubAssemblyTag(size_t _subId) const
 {
-	assertThrow(m_data < (u256(1) << 64), Exception, "Tag already has subassembly set.");
+	assertThrow(data() < (u256(1) << 64), Exception, "Tag already has subassembly set.");
 
 	assertThrow(m_type == PushTag || m_type == Tag, Exception, "");
 	AssemblyItem r = *this;
 	r.m_type = PushTag;
-	r.setPushTagSubIdAndTag(_subId, size_t(m_data));
+	r.setPushTagSubIdAndTag(_subId, size_t(data()));
 	return r;
 }
 
 pair<size_t, size_t> AssemblyItem::splitForeignPushTag() const
 {
 	assertThrow(m_type == PushTag || m_type == Tag, Exception, "");
-	return make_pair(size_t(m_data / (u256(1) << 64)) - 1, size_t(m_data));
+	return make_pair(size_t((data()) / (u256(1) << 64)) - 1, size_t(data()));
 }
 
 void AssemblyItem::setPushTagSubIdAndTag(size_t _subId, size_t _tag)
@@ -59,7 +60,7 @@ unsigned AssemblyItem::bytesRequired(unsigned _addressLength) const
 	case PushString:
 		return 33;
 	case Push:
-		return 1 + max<unsigned>(1, dev::bytesRequired(m_data));
+		return 1 + max<unsigned>(1, dev::bytesRequired(data()));
 	case PushSubSize:
 	case PushProgramSize:
 		return 4;		// worst case: a 16MB program
@@ -97,6 +98,28 @@ int AssemblyItem::deposit() const
 	return 0;
 }
 
+bool AssemblyItem::canBeFunctional() const
+{
+	switch (m_type)
+	{
+	case Operation:
+		return !SemanticInformation::isDupInstruction(*this) && !SemanticInformation::isSwapInstruction(*this);
+	case Push:
+	case PushString:
+	case PushTag:
+	case PushData:
+	case PushSub:
+	case PushSubSize:
+	case PushProgramSize:
+	case PushLibraryAddress:
+		return true;
+	case Tag:
+		return false;
+	default:;
+	}
+	return 0;
+}
+
 string AssemblyItem::getJumpTypeAsString() const
 {
 	switch (m_jumpType)
@@ -109,6 +132,65 @@ string AssemblyItem::getJumpTypeAsString() const
 	default:
 		return "";
 	}
+}
+
+string AssemblyItem::toAssemblyText() const
+{
+	string text;
+	switch (type())
+	{
+	case Operation:
+	{
+		assertThrow(isValidInstruction(instruction()), AssemblyException, "Invalid instruction.");
+		string name = instructionInfo(instruction()).name;
+		transform(name.begin(), name.end(), name.begin(), [](unsigned char _c) { return tolower(_c); });
+		text = name;
+		break;
+	}
+	case Push:
+		text = toHex(toCompactBigEndian(data(), 1), 1, HexPrefix::Add);
+		break;
+	case PushString:
+		assertThrow(false, AssemblyException, "Push string assembly output not implemented.");
+		break;
+	case PushTag:
+		assertThrow(data() < 0x10000, AssemblyException, "Sub-assembly tags not yet implemented.");
+		text = string("tag_") + to_string(size_t(data()));
+		break;
+	case Tag:
+		assertThrow(data() < 0x10000, AssemblyException, "Sub-assembly tags not yet implemented.");
+		text = string("tag_") + to_string(size_t(data())) + ":";
+		break;
+	case PushData:
+		assertThrow(false, AssemblyException, "Push data not implemented.");
+		break;
+	case PushSub:
+		text = string("dataOffset(sub_") + to_string(size_t(data())) + ")";
+		break;
+	case PushSubSize:
+		text = string("dataSize(sub_") + to_string(size_t(data())) + ")";
+		break;
+	case PushProgramSize:
+		text = string("bytecodeSize");
+		break;
+	case PushLibraryAddress:
+		text = string("linkerSymbol(\"") + toHex(data()) + string("\")");
+		break;
+	case UndefinedItem:
+		assertThrow(false, AssemblyException, "Invalid assembly item.");
+		break;
+	default:
+		BOOST_THROW_EXCEPTION(InvalidOpcode());
+	}
+	if (m_jumpType == JumpType::IntoFunction || m_jumpType == JumpType::OutOfFunction)
+	{
+		text += "\t//";
+		if (m_jumpType == JumpType::IntoFunction)
+			text += " in";
+		else
+			text += " out";
+	}
+	return text;
 }
 
 ostream& dev::eth::operator<<(ostream& _out, AssemblyItem const& _item)